requirements.txt
## lista de librerías necesarias para el proyecto




opencv-python==4.8.0.76
pytesseract == 0.3.10
Pillow == 9.5.0
numpy == 1.24.3
pandas == 1.5.3
##tensorflow== 2.12.0
scikit-learn== 1.2.2
matplotlib== 3.7.1
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0


